services/inference_service/inference_requirements.txt
grpcio
grpcio-tools
openai
pinecone-client